IS_Student_AI_Assistant/requirements.txt
customtkinter>=5.2.2
SQLAlchemy>=2.0.30
python-dotenv>=1.0.1
google-genai>=0.7.0